Fix test setup
