+++ app/controllers/history_controller.rb
@@ -85,4 +85,3 @@ class HistoryController < ApplicationController
 	end
 
 end
-
