+++ config/environments/test.rb
@@ -1,35 +1,36 @@
 Rails.application.configure do
+  # Settings specified here will take precedence over those in config/environment.rb
 
-# Settings specified here will take precedence over those in config/environment.rb
+  # The test environment is used exclusively to run your application's
+  # test suite.  You never need to work with it otherwise.  Remember that
+  # your test database is "scratch space" for the test suite and is wiped
+  # and recreated between test runs.  Don't rely on the data there!
+  config.cache_classes = true
 
-# The test environment is used exclusively to run your application's
-# test suite.  You never need to work with it otherwise.  Remember that
-# your test database is "scratch space" for the test suite and is wiped
-# and recreated between test runs.  Don't rely on the data there!
-config.cache_classes = true
+  # Log error messages when you accidentally call methods on nil.
+  config.whiny_nils = true
 
-# Log error messages when you accidentally call methods on nil.
-config.whiny_nils = true
+  # Show full error reports and disable caching
+  #error config.action_controller.consider_all_requests_local = true
+  config.action_controller.perform_caching             = false
+  config.action_view.cache_template_loading            = true
+  config.eager_load=false
+  # Disable request forgery protection in test environment
+  config.action_controller.allow_forgery_protection    = false
 
-# Show full error reports and disable caching
-#error config.action_controller.consider_all_requests_local = true
-config.action_controller.perform_caching             = false
-config.action_view.cache_template_loading            = true
-config.eager_load=false
-# Disable request forgery protection in test environment
-config.action_controller.allow_forgery_protection    = false
+  # Tell Action Mailer not to deliver emails to the real world.
+  # The :test delivery method accumulates sent emails in the
+  # ActionMailer::Base.deliveries array.
+  config.action_mailer.delivery_method = :test
+  config.action_mailer.default_url_options = {
+      :host => "localhost:3000"
+    }
+  # Use SQL instead of Active Record's schema dumper when creating the test database.
+  # This is necessary if your schema can't be completely dumped by the schema dumper,
+  # like if you have constraints or database-specific column types
+  # config.active_record.schema_format = :sql
 
-# Tell Action Mailer not to deliver emails to the real world.
-# The :test delivery method accumulates sent emails in the
-# ActionMailer::Base.deliveries array.
-config.action_mailer.delivery_method = :test
-config.action_mailer.default_url_options = {
-    :host => "localhost:3000"
-  }
-# Use SQL instead of Active Record's schema dumper when creating the test database.
-# This is necessary if your schema can't be completely dumped by the schema dumper,
-# like if you have constraints or database-specific column types
-# config.active_record.schema_format = :sql
+  config.log_level = :error
 
-config.log_level = :debug
+  config.active_support.test_order = :random
 end
