+++ config/initializers/sylrplm_init.rb
@@ -95,4 +95,3 @@ if (false)
 	#Rake::Task["notify"].invoke
 	end
 end
-
