+++ lib/controllers/plm_object_controller.rb
