+++ lib/models/plm_object.rb
