+++ lib/ruote/sylrplm/history_entry.rb
@@ -1,82 +1,103 @@
-require 'openwfe/extras/expool/db_history'
-
 module Ruote
-	module Sylrplm
-		class HistoryEntry < OpenWFE::Extras::HistoryEntry
-			#include Models::PlmObject
-			include Models::SylrplmCommon
+  module Sylrplm
+    #
+    # The active record for process errors. Not much to say.
+    #
+    class HistoryEntry < ActiveRecord::Base
+      include OpenWFE::Extras::SingleConnectionMixin
+      include Models::SylrplmCommon
 
-			#rails2 has_many :links_customers, :class_name => "Link", :foreign_key => "child_id", :conditions => ["father_plmtype='history_entry' and child_plmtype='customer'"]
-			has_many :links_customers, -> { where(father_plmtype: 'history_entry'  , child_plmtype: 'customer') }, :class_name => "Link", :foreign_key => "child_id"
+      attr_reader :link_attributes
 
-			has_many :customers , :through => :links_customers, :source => :customer_down
+      #TODO syl set_table_name('history')
+      #ko rails 4set_table_name('history_entry')
+      #ok rails4
+      self.table_name = 'history_entry'
 
-			#rails2 has_many :links_documents, :class_name => "Link", :foreign_key => "child_id", :conditions => ["father_plmtype='history_entry' and child_plmtype='document'"]
-			has_many :links_documents, -> { where(father_plmtype: 'history_entry'  , child_plmtype: 'document') }, :class_name => "Link", :foreign_key => "child_id"
+      def self.model_name
+        "history_entry"
+      end
 
-			has_many :documents , :through => :links_documents, :source => :document_down
+      #
+      # returns a FlowExpressionId instance if the entry has a 'fei' or
+      # nil instead.
+      #
+      def full_fei
+        self.fei ? OpenWFE::FlowExpressionId.from_s(self.fei) : nil
+      end
 
-			#rails2 has_many :links_parts, :class_name => "Link", :foreign_key => "child_id", :conditions => ["father_plmtype='history_entry' and child_plmtype='part'"]
-			has_many :links_parts, -> { where(father_plmtype: 'history_entry'  , child_plmtype: 'part') }, :class_name => "Link", :foreign_key => "child_id"
-			has_many :parts , :through => :links_parts, :source => :part_down
+      #
+      # Directly logs an event (may throw an exception)
+      #
+      def self.log! (source, event, opts={})
+        fei = opts[:fei]
 
-			#rails2 has_many :links_projects, :class_name => "Link", :foreign_key => "child_id", :conditions => ["father_plmtype='history_entry' and child_plmtype='project'"]
-			has_many :links_projects, -> { where(father_plmtype: 'history_entry'  , child_plmtype: 'project') }, :class_name => "Link", :foreign_key => "child_id"
+        if fei
+          opts[:wfid] = fei.parent_wfid
+          opts[:wfname] = fei.wfname
+          opts[:wfrevision] = fei.wfrevision
+          opts[:fei] = fei.to_s
+        end
 
-			has_many :projects , :through => :links_projects, :source => :project_down
+        opts[:source] = source.to_s
+        opts[:event] = event.to_s
 
-			#rails2 has_many :links_plmobjects, :class_name => "Link", :foreign_key => "child_id", :conditions => ["father_plmtype='history_entry'"]
-			has_many :links_plmobjects, -> { where(father_plmtype: 'history_entry' ) }, :class_name => "Link", :foreign_key => "child_id"
+        #self.new(opts).save!
+        # syl: unknown attribute
+        #puts "db_history.log:opts=#{opts.inspect}"
+        opts.delete(:inflow)
+        ret=self.new(opts)
+        st=ret.save_without_transactions!
+        #begin
+        #  self.new(opts).save!
+        #rescue Exception => e
+        #  puts ; puts e
+        #  self.new(opts).save! rescue nil
+        #end
+        #_syl_ pour recuperer l'entry
+        #puts "********** db_history.log:wi_fields=#{ret.fields}"
+        (st ? ret : nil)
+      end
 
-			attr_accessor :link_attributes
+      attr_accessor :link_attributes
 
-			def plm_objects
-				fname= "#{self.class.name}.#{__method__}"
-				ret=[]
-				#self.links_plmobjects.each do |lnk|
-				cond = "father_plmtype='history_entry' and father_id = '#{self.id}'"
-				lnks = Link.find(:all, :conditions => [cond])
-				#LOG.debug (fname){"cond=#{cond} #{lnks.count} link trouves"}
-				lnks.each  do |lnk|
-					#LOG.debug (fname){"lnk=#{lnk} father=#{lnk.father.ident} child=#{lnk.child.ident}"}
-					ret.push lnk.child unless lnk.child.nil?
-				end
-				ret
-			end
+      def plm_objects
+        fname= "#{self.class.name}.#{__method__}"
+        ret=[]
+        #self.links_plmobjects.each do |lnk|
+        cond = "father_plmtype='history_entry' and father_id = '#{self.id}'"
+        lnks = Link.find(:all, :conditions => [cond])
+        #LOG.debug (fname){"cond=#{cond} #{lnks.count} link trouves"}
+        lnks.each do |lnk|
+          #LOG.debug (fname){"lnk=#{lnk} father=#{lnk.father.ident} child=#{lnk.child.ident}"}
+          ret.push lnk.child unless lnk.child.nil?
+        end
+        ret
+      end
 
-			def link_attributes=(att)
-				fname= "#{self.class.name}.#{__method__}"
-				@link_attributes = att
-				#LOG.debug (fname){"HistoryEntry:link_attributes=#{@link_attributes}"}
-				@link_attributes
-			end
+      def link_attributes=(att)
+        fname= "#{self.class.name}.#{__method__}"
+        @link_attributes = att
+        #LOG.debug (fname){"HistoryEntry:link_attributes=#{@link_attributes}"}
+        @link_attributes
+      end
 
-			def link_attributes
-				@link_attributes
-			end
+      def typesobject
+        Typesobject.find_by_forobject(model_name).to_a[0]
+      end
 
-			def typesobject
-				Typesobject.find_by_forobject(model_name).to_a[0]
-			end
+      def model_name
+        self.class.model_name
+      end
 
-			def model_name
-				HistoryEntry.model_name
-			end
+      def ident
+        #fei.to_s+"_"+wfid.to_s+"_"+wf_name.to_s
+        [wfid, wf_name].join("_")
+      end
 
-			def self.model_name
-				"history_entry"
-			end
-
-			def ident
-				#fei.to_s+"_"+wfid.to_s+"_"+wf_name.to_s
-				[wfid, wf_name].join("_")
-			end
-
-			def cancel?
-				source=='expool' && event=='cancel'
-			end
-
-		end
-	end
-
+      def cancel?
+        source=='expool' && event=='cancel'
+      end
+    end
+  end
 end
