+++ lib/ruote/sylrplm/sylrplm.rb
@@ -5,6 +5,3 @@ require 'ruote/sylrplm/plm_participant'
 require 'ruote/sylrplm/plm_process_exception'
 require 'ruote/sylrplm/process'
 require 'ruote/sylrplm/singlecon'
-
-
-
