+++ lib/tasks/sylrplm_tasks.rake
@@ -481,4 +481,3 @@ namespace :sylrplm do
 	end
 
 end
-
